break out mime types
